LayoutTests/imported/w3c/web-platform-tests/webxr/gamepads-module/xrInputSource_gamepad_disconnect.https-expected.txt


FAIL WebXR InputSource's gamepad gets disconnected when the input source is removed - webgl assert_equals: WebXR Gamepad.index must be -1, iteration: 1 expected -1 but got 4294967295
FAIL WebXR InputSource's gamepad gets disconnected when the input source is removed - webgl2 promise_test: Unhandled rejection with value: "Session with params \"immersive-vr\" was rejected on device {\"supportsImmersive\":true,\"supportedModes\":[\"inline\",\"immersive-vr\"],\"views\":[{\"eye\":\"left\",\"projectionMatrix\":[1,0,0,0,0,1,0,0,3,2,-1,-1,0,0,-0.2,0],\"viewOffset\":{\"position\":[-0.1,0,0],\"orientation\":[0,0,0,1]},\"resolution\":{\"width\":200,\"height\":200}},{\"eye\":\"right\",\"projectionMatrix\":[1,0,0,0,0,1,0,0,3,2,-1,-1,0,0,-0.2,0],\"viewOffset\":{\"position\":[0.1,0,0],\"orientation\":[0,0,0,1]},\"resolution\":{\"width\":200,\"height\":200}}],\"viewerOrigin\":{\"position\":[0,0,0],\"orientation\":[0,0,0,1]},\"supportedFeatures\":[\"viewer\",\"local\",\"local-floor\",\"bounded-floor\",\"unbounded\",\"hit-test\",\"dom-overlay\",\"light-estimation\",\"anchors\"],\"environmentBlendMode\":\"opaque\",\"interactionMode\":\"world-space\"} with error: OperationError: Unable to create a framebuffer."
